\name{team05Lab03-package}
\alias{team05Lab03-package}
\alias{team05Lab03}
\docType{package}
\title{
\packageTitle{team05Lab03}
}
\description{
\packageDescription{team05Lab03}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{team05Lab03}
\packageIndices{team05Lab03}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{team05Lab03}

Maintainer: \packageMaintainer{team05Lab03}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
